// src/models/Teacher.ts
/* Archivo: Frontend\src\models\Teacher.ts
   Proposito: Implementa la logica principal del archivo Teacher.
*/
export interface Teacher {
  id?: string;
  user_id: string;
  first_name: string;
  last_name: string;
  phone?: string;
  identification: string;
  specialty?: string;
  is_active?: boolean;
}